add symlink

+++ macos/files/.bin/Symlinks.sh
@@ -25,18 +25,39 @@ fi
 done
 
 ## .ssh
-if [ -d ${HOME}/.repos/bitbucket.org/libero18/.ssh ]; then
-  ln -fs $HOME/.repos/bitbucket.org/libero18/.ssh $HOME >/dev/null 2>&1
-fi
-
-## anyenv
-if [ -d ${HOME}/.repos/github.com/riywo/anyenv ]; then
-  ln -fs $HOME/.repos/github.com/riywo/anyenv $HOME/.anyenv >/dev/null 2>&1
+if [ -d ${HOME}/.repos/bitbucket.org/libero18/.ssh -a ! -L $HOME/.ssh ]; then
+  ln -fs $HOME/.repos/bitbucket.org/libero18/.ssh $HOME/.ssh >/dev/null 2>&1
 fi
 
 ## oh-my-zsh theme
 if [ ! -L ${HOME}/.repos/github.com/robbyrussell/oh-my-zsh/custom/libero18.zsh-theme ]; then
   ln -fs ${HOME}/.repos/github.com/libero18/dotfiles/macos/files/.etc/libero18.zsh-theme ${HOME}/repos/github.com/robbyrussell/oh-my-zsh/custom/ >/dev/null 2>&1
+fi
+
+## anyenv
+if [ -d ${HOME}/.repos/github.com/riywo/anyenv -a ! -L $HOME/.anyenv ]; then
+  ln -fs $HOME/.repos/github.com/riywo/anyenv $HOME/.anyenv >/dev/null 2>&1
+fi
+
+## anyenv-update
+if [ -L $HOME/.anyenv ]; then
+  if [ ! -d $HOME/.anyenv/plugins ]; then
+    mkdir -p $HOME/.anyenv/plugins
+  fi
+  if [ ! -L $HOME/.anyenv/plugins/anyenv-update ]; then
+    ln -fs $HOME/.repos/github.com/znz/anyenv-update $HOME/.anyenv/plugins/anyenv-update >/dev/null 2>&1
+  fi
+fi
+
+## rbenv-plug
+if [ -L $HOME/.anyenv ]; then
+  if [ ! -d ${HOME}/.anyenv/envs/rbenv ]; then
+    ${HOME}/.anyenv/bin/anyenv install rbenv
+    source $HOME/.zshrc
+  fi
+  if [ ! -L $HOME/.anyenv/envs/rbenv/plugins/rbenv-plug ]; then
+    ln -fs $HOME/.repos/github.com/znz/rbenv-plug $HOME/.anyenv/envs/rbenv/plugins/rbenv-plug >/dev/null 2>&1
+  fi
 fi
 
 cd - >/dev/null 2>&1
